requirements.txt
# Core Streamlit and web framework
streamlit==1.28.1
streamlit-authenticator==0.2.3

# Supabase for authentication and database
supabase==2.0.0
postgrest==0.13.0

# LangChain and AI components
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.6
openai==1.3.5
tiktoken==0.5.1

# Vector database and embeddings
chromadb==0.4.15
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Document processing
pypdf2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
pillow==10.1.0

# Multilingual support
googletrans==4.0.0rc1
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3

# Task scheduling and background jobs
schedule==1.2.0
celery==5.3.4

# Date and time handling
datetime==5.3
pytz==2023.3

# Logging and monitoring
loguru==0.7.2

# Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0